codes/backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
click==8.3.0
distro==1.9.0
fastapi==0.115.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
idna==3.11
jiter==0.12.0
openai==2.8.0
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.0.1
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.38.6
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.30.6
watchfiles==1.1.1
websockets==15.0.1
requests==2.32.3
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2
tensorflow==2.16.1
